refactor: re-structure job_list to job/list
