+++ apps/web/src/modules/job/detail/components/JobCard/components/JobCardFooter/styled.ts
