+++ apps/web/src/modules/job/detail/components/JobCard/components/JobCardHeader/styled.ts
