+++ apps/web/src/modules/job/detail/components/JobCard/components/ShootingDetail/index.tsx
