+++ apps/web/src/modules/job/detail/components/JobCard/components/ShootingDetail/type.ts
