+++ apps/web/src/modules/job/detail/components/JobCard/styled.ts
