+++ apps/web/src/modules/job/detail/pages/styled.ts
