+++ apps/web/src/modules/job/list/components/FilterContainer/index.tsx
@@ -12,7 +12,7 @@ import {
 } from '@mui/material'
 import React from 'react'
 
-import TextField from '../../../../common/components/TextField'
+import TextField from '../../../../../common/components/TextField'
 import { FilterBox } from './style'
 
 export default function SearchBox() {
