+++ apps/web/src/modules/job/list/components/JobCardContainer/styled.ts
