+++ apps/web/src/modules/job/list/components/NotiCardContainer/components/NotiCard/components/Header/styled.ts
