+++ apps/web/src/modules/job/list/components/NotiCardContainer/components/NotiCard/components/OfferDetail/styled.ts
