+++ apps/web/src/modules/job/list/components/NotiCardContainer/components/NotiCard/components/OfferDetail/type.ts
