+++ apps/web/src/modules/job/list/components/NotiCardContainer/components/NotiCard/styled.ts
