+++ apps/web/src/modules/job/list/components/NotiCardContainer/styled.ts
