+++ apps/web/src/modules/job/list/components/NotiCardContainer/type.ts
