+++ apps/web/src/modules/job/list/components/SearchBox/index.tsx
@@ -1,7 +1,7 @@
 import SearchOutlinedIcon from '@mui/icons-material/SearchOutlined'
 import React from 'react'
 
-import TextField from '../../../../common/components/TextField'
+import TextField from '../../../../../common/components/TextField'
 
 export default function SearchBox() {
   return (
