+++ apps/web/src/modules/job/list/pages/styled.ts
