+++ apps/web/src/pages/job.tsx
@@ -1,3 +1,3 @@
-import JobListPage from 'modules/job_list/pages'
+import JobListPage from 'modules/job/list/pages'
 
 export default JobListPage
